don't write the index if index.html already exists

+++ src/tfwebserver/sitetypes.cr
@@ -11,17 +11,16 @@ module TFWeb
     property autocommit = false
     property srcdir = "src"
     property environment = ""
-    property jinja_env = Crinja.new()
+    property jinja_env = Crinja.new
 
     # def initialize()
     #   @jinja_env = Crinja.new()
     # end
-    
+
     def prepare_on_fs
       repo = self.repo
       templatesdir = File.join(@path, @srcdir, "templates")
       if Dir.exists?(@path)
-        puts "TEMPLATE LOADED FOR #{templatesdir}"
         @jinja_env.loader = Crinja::Loader::FileSystemLoader.new(templatesdir)
       end
     end
@@ -35,15 +34,16 @@ module TFWeb
     end
   end
 
-  class Wiki < Site   
-
+  class Wiki < Site
     private def prepare_index
       repo = self.repo
-      title =  @title.size > 0 ? @title : @name
+      title = @title.size > 0 ? @title : @name
       url_as_https = repo.not_nil!.url_as_https || ""
       html = render "src/tfwebserver/views/docsify.ecr"
       destindex = File.join(@path, @srcdir, "index.html")
-      File.write(destindex, html)
+      unless File.exists?(destindex)
+        File.write(destindex, html)
+      end
     end
 
     def prepare_on_fs
